requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
Jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.1
weasyprint==62.3
